Moved removeValidator to ssv.js so registerValidators Action doesn't need the SSV packages signers.js removed dependency on test utility
